Start work on the inspect-button feature
I got the copying of the inspectElement html code
done, I just need to try making an "inspect screen"
that shows off some of the potential for items.
Ooo a toggle for the on off broadcast would be really
cool. I've just decided that I want an equip button
if you designate an item as equippable. It makes a
button trigger the onclick event for the item.